requirements.txt
# Core dependencies for the Gemini Financial Chatbot

google-generativeai>=0.5.4
pandas>=2.2.0
openpyxl>=3.1.0
python-docx>=1.1.0 